% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/why-source.R
\name{compute_pca_from_features}
\alias{compute_pca_from_features}
\title{PCA from features}
\usage{
compute_pca_from_features(
  feats_folder,
  otp,
  ftp,
  axis_x,
  axis_y,
  color_by_SE_vars,
  SE_data_file,
  get_point_identity
)
}
\arguments{
\item{feats_folder}{String with the path to the folder where the features are contained.}

\item{otp}{Vector of the observations to plot.}

\item{ftp}{Vector of the features to plot.}

\item{axis_x}{Integer indicating the principal component to be plotted as axis x.}

\item{axis_y}{Integer indicating the principal component to be plotted as axis y.}

\item{color_by_SE_vars}{Boolean indicating if plotted points must be colored according to the socioeconomical variables.}

\item{SE_data_file}{String with the path to the file with the socioeconomical variables.}

\item{get_point_identity}{Boolean indicating if plotted points must be identified by clicking on them.}
}
\value{
A list with class `princomp`.
}
\description{
Compute and plot PCA from the features of the time series.
}
