% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/why-source.R
\name{dataset_to_raw_dataframe}
\alias{dataset_to_raw_dataframe}
\title{Raw dataframe from dataset}
\usage{
dataset_to_raw_dataframe(csv_file)
}
\arguments{
\item{csv_file}{Path (string) to the CSV file containing the dataset.}
}
\value{
Raw dataframe, i.e. without checking missing samples.
}
\description{
Get a dataframe from a dataset contained in a CSV file.
}
