% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/why-source.R
\name{electricity}
\alias{electricity}
\title{Features of electricity}
\usage{
electricity(x)
}
\arguments{
\item{x}{Time series of class `msts`.}
}
\value{
A list with the mean and variance of the load factors across days in the time series.
}
\description{
Compute the load factor across days of a time series.
}
