% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/why-source.R
\name{Get_Data_Interval}
\alias{Get_Data_Interval}
\title{Extraction of a time series according to a time interval}
\usage{
Get_Data_Interval(dset_data, from_date, to_date, sampling_period)
}
\arguments{
\item{dset_data}{Time series in the proper data.frame format.}

\item{from_date}{The initial GMT date of the interval.}

\item{to_date}{The final GMT date of the interval.}

\item{sampling_period}{Sampling period of the data.frame.}
}
\value{
A data.frame containing GMT dates and values.
}
\description{
Extract a time series contained between two dates.
}
\details{
`sampling_period` (measured in seconds) is needed as an input parameter to check non-existing dates, which must be indicated as `NaN`.
}
