% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-features-from-cooked-dataframe.R
\name{get_features_from_cooked_dataframe}
\alias{get_features_from_cooked_dataframe}
\title{Features of a cooked dataframe}
\usage{
get_features_from_cooked_dataframe(df, type_of_analysis)
}
\arguments{
\item{type_of_analysis}{A string indicating the type of analysis: either `basic` or `extra`. `basic` contains 7 functions whereas `extra` contains 33.}

\item{cooked_df}{Cooked dataframe.}
}
\value{
List of features.
}
\description{
Get features of a cooked dataframe.
}
