% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-raw-dataframe-from-dataset.R
\name{get_raw_dataframe_from_dataset}
\alias{get_raw_dataframe_from_dataset}
\title{Raw dataframe from dataset}
\usage{
get_raw_dataframe_from_dataset(csv_file)
}
\arguments{
\item{csv_file}{String with the absolute path to the CSV file containing the dataset.}
}
\value{
Raw dataframe, i.e. without checking missing samples.
}
\description{
Get a dataframe from a dataset contained in a CSV file.
}
