% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/why-source.R
\name{Load_Dataset_File}
\alias{Load_Dataset_File}
\title{Loading of time series}
\usage{
Load_Dataset_File(dataset_path, filename)
}
\arguments{
\item{dataset_path}{Path to the proper CSV file.}

\item{filename}{Name of the proper CSV file.}
}
\value{
A data.frame containing GMT dates and values.
}
\description{
Load a time series contained in a proper CSV file.
}
