% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/why-source.R
\name{Plot_Data}
\alias{Plot_Data}
\title{Plot of a time series contained in a data.frame}
\usage{
Plot_Data(dset_data, title = NULL)
}
\arguments{
\item{dset_data}{The data.frame to be plotted.}

\item{title}{Optional title (`NULL` by default).}
}
\value{
Plot of the time series.
}
\description{
Plot a time series contained in a proper data.frame.
}
