% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/why-source.R
\name{Plot_Data_Info_Row}
\alias{Plot_Data_Info_Row}
\title{Plot of a row of `data_info`}
\usage{
Plot_Data_Info_Row(di_row, dset_key = "lcl")
}
\arguments{
\item{di_row}{Row of `data_info`.}

\item{dset_key}{Key of the dataset.}
}
\value{
Plot of the time series.
}
\description{
Plot the time series contained in a row of the file `data_info.csv` (once loaded).
}
