% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/why-source.R
\name{plot_dataframe}
\alias{plot_dataframe}
\title{Plotting of a dataframe}
\usage{
plot_dataframe(dset_data, title = NULL)
}
\arguments{
\item{dset_data}{The dataframe to be plotted.}

\item{title}{Optional title (`NULL` by default).}
}
\value{
Plot of the dataset.
}
\description{
Plot a dataframe (either raw or cooked).
}
