% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/why-source.R
\name{plot_datainfo_row}
\alias{plot_datainfo_row}
\title{Plotting of a row of `data_info` plotting}
\usage{
plot_datainfo_row(di_row, dset_key = "lcl")
}
\arguments{
\item{di_row}{Row of `data_info`.}

\item{dset_key}{Key of the dataset.}
}
\value{
Plot of the dataset.
}
\description{
Plot the dataset indicated in a row of the file `data_info.csv`.
}
