% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/why-source.R
\name{plot_features_library}
\alias{plot_features_library}
\title{Creation of a PDF library of features}
\usage{
plot_features_library(sampling_period, feats_folder, feats_to_plot)
}
\arguments{
\item{sampling_period}{Sampling period of the dataframe.}

\item{feats_folder}{String with a path to the folder where the features are contained.}

\item{feats_to_plot}{Vector of the features to be plotted.}
}
\value{
A PDF file per feature with nine plots each.
}
\description{
For each feature, the nine most representative values are plotted, including the minimum, the median and the maximum values.
}
