% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/why-source.R
\name{raw_to_cooked_dataframe}
\alias{raw_to_cooked_dataframe}
\title{Cooked dataframe from raw dataframe}
\usage{
raw_to_cooked_dataframe(dset_data, from_date, to_date, dset_key)
}
\arguments{
\item{dset_data}{Raw dataframe.}

\item{from_date}{Initial date and time of the interval. Either a `POSIXct` class in the GMT time zone OR a string `first`.}

\item{to_date}{Final date and time of the interval. Either a `POSIXct` class in the GMT time zone OR a string `last`.}

\item{dset_key}{String indicating the key of the dataset. `lcl` for `Low Carbon London`.}
}
\value{
List with the cooked dataframe, a vector of seasonal periods, a double with the percentage of `NA` values, and a boolean indicating if the dataframe values are all zeros.
}
\description{
Cook a raw dataframe. Cooking consists in completing missing samples with NA values, removing extra samples (those not matching the sampling period), extracting a user-defined time interval out of the raw dataframe, and checking its validity to be feature-analyzed.
}
