% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/why-source.R
\name{raw_to_cooked_dataframe}
\alias{raw_to_cooked_dataframe}
\title{Cooked dataframe from raw dataframe}
\usage{
raw_to_cooked_dataframe(dset_data, from_date, to_date, sampling_period)
}
\arguments{
\item{dset_data}{Raw dataframe.}

\item{from_date}{Initial date and time of the interval of class `POSIXct` and time zone GMT.}

\item{to_date}{Final date and time of the interval of class `POSIXct` and time zone GMT.}

\item{sampling_period}{Sampling period (measured in seconds) of the raw dataframe. This value is required to complete missing samples with `NaN`.}
}
\value{
Cooked dataframe.
}
\description{
Cook a raw dataframe. Cooking consists in completing missing samples with NA values and extracting a time interval out of the raw dataframe.
}
